Add tests for query function
