+++ tests/Nhs.Appointments.Api.Integration/Scenarios/CreateAvailability/ApplyAvailabilityTemplate.feature
@@ -10,7 +10,7 @@
       | Tomorrow    | 09:00 | 10:00 | COVID    | 5           | 1        |
       | Tomorrow_+1 | 09:00 | 10:00 | COVID    | 5           | 1        |
       | Tomorrow_+2 | 09:00 | 10:00 | COVID    | 5           | 1        |
-    And the following availability created event is created
+    And the following availability created events are created
       | Type     | By       | FromDate | ToDate      | Template_Days | FromTime | UntilTime | SlotLength | Capacity | Services |
       | Template | api@test | Tomorrow | Tomorrow_+2 | Relative      | 09:00    | 10:00     | 5          | 1        | COVID    |
