+++ tests/Nhs.Appointments.Api.Integration/Scenarios/CreateAvailability/GetAvailabilityCreatedEventsFeatureSteps.cs
@@ -1,0 +1,33 @@
+﻿using System.Collections.Generic;
+using System.Net;
+using System.Net.Http;
+using System.Threading.Tasks;
+using FluentAssertions;
+using Nhs.Appointments.Api.Json;
+using Xunit.Gherkin.Quick;
+using Nhs.Appointments.Api.Models;
+using Nhs.Appointments.Core;
+
+namespace Nhs.Appointments.Api.Integration.Scenarios.CreateAvailability
+{
+    [FeatureFile("./Scenarios/CreateAvailability/GetAvailabilityCreatedEvents.feature")]
+    public sealed class GetAvailabilityCreatedEventsFeatureSteps : BaseCreateAvailabilityFeatureSteps
+    {
+        private HttpResponseMessage _response;
+        private HttpStatusCode _statusCode;
+        private IEnumerable<AvailabilityCreatedEvent> _actualResponse;
+
+        [Then("I request Availability Created Events for the current site")]
+        public async Task ThenIRequestAvailabilityCreatedEventsForTheCurrentSite()
+        {
+            var siteId = GetSiteId();
+
+            _response = await Http.GetAsync($"http://localhost:7071/api/availability-created?site={siteId}");
+            _statusCode = _response.StatusCode;
+            _actualResponse = await JsonRequestReader.ReadRequestAsync<IEnumerable<AvailabilityCreatedEvent>>(await _response.Content.ReadAsStreamAsync());
+
+            _statusCode.Should().Be(HttpStatusCode.OK);
+            _expectedAvailabilityCreatedEvents.Should().BeEquivalentTo(_actualResponse);
+        }
+    }
+}
