+++ tests/Nhs.Appointments.Api.Integration/Scenarios/CreateAvailability/SetAvailability.feature
@@ -8,6 +8,6 @@
     Then the request is successful and the following availability is created
       | Date     | From  | Until | Services | Slot Length | Capacity |
       | Tomorrow | 09:00 | 17:00 | COVID    | 5           | 1        |
-    And the following availability created event is created
+    And the following availability created events are created
       | Type              | By       | FromDate | ToDate | Template_Days | FromTime | UntilTime | SlotLength | Capacity | Services |
       | SingleDateSession | api@test | Tomorrow |        |               | 09:00    | 17:00     | 5          | 1        | COVID    |
