+++ tests/Nhs.Appointments.Api.UnitTests/Functions/GetAvailabilityCreatedEventsFunctionTests.cs
